Call handler immediately when triggered
Previously, the infinite scroll handler wasn’t called until when the
scrollview was about to stop scrolling, that is, when the run loop mode
exited UITrackingRunLoopMode and switched to NSDefaultRunLoopMode. When
using the infiniteScrollTriggerOffset, this caused the handler to be
triggered not when the threshold was crossed, but when scrolling was
finished. Calling the handler instantly fixes this. However there may
be a good reason it was originally set to be executed after scrolling
(at the moment I don’t know what that could be).